requirements.txt
nextcord
python-dotenv
aiofiles
google-api-core
google-auth
google-generativeai
aiosqlite
openai
requests
aiohttp
Pillow